Some minor modifications

+++ drop_db.sh
@@ -3,7 +3,7 @@
 while true; do
     read -r -p "Enter DB Name: " name
 
-    if [[ -d "./DB/$name" && -e "./DB/$name" ]]; then
+    if [[ -n "$name" && -d "./DB/$name" && -e "./DB/$name" ]]; then
         rm -r "./DB/$name"
         echo -e "\e[32mSuccess: Database '$name' deleted.\e[0m"
         break
